pytest: split kcm test cases in to separate file.
